requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.14.0
groq==0.4.1
pydantic==2.5.3
